# storyboard_studio/data/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Storyboard Studio API",
    "description": "A local-first, editable PowerPoint presentation generator.",
    "version": "1.1.0",
    "x-storyboard-contract": {
      "schema": "storyboard-v1",
      "story_schema": "story-v2",
      "compatibility": "docs/SUPPORT_MATRIX.md",
      "migrations": "docs/MIGRATIONS.md"
    }
  },
  "paths": {
...
    "/api/v1/layout/preflight": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Layout Preflight",
        "description": "Find deterministic text overflow risks before creating a PowerPoint.",
        "operationId": "layout_preflight_api_v1_layout_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Layout Preflight Api V1 Layout Preflight Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/coverage": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Evidence Coverage Report",
        "description": "Map claims to explicit author-linked and author-checked sources.",
        "operationId": "evidence_coverage_report_api_v1_evidence_coverage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evidence Coverage Report Api V1 Evidence Coverage Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctor": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Doctor",
        "description": "Diagnose narrative structure and evidence gaps without a network provider.",
        "operationId": "doctor_api_v1_doctor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Doctor Api V1 Doctor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/doctor": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Story Doctor",
        "description": "Diagnose a versioned story and preserve explicit finding dispositions.",
        "operationId": "story_doctor_api_v1_stories_doctor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryDocumentV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Story Doctor Api V1 Stories Doctor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stories/decision-brief": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Create Decision Story",
        "description": "Compile an author-supplied decision brief locally without invented claims.",
        "operationId": "create_decision_story_api_v1_stories_decision_brief_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuidedDecisionRequest"
              },
              "examples": {
                "guided-decision": {
                  "summary": "Guided Decision",
                  "value": {
                    "brief": {
                      "schema_version": "2",
                      "template": "decision-brief",
                      "decision": "Choose the onboarding pilot for the next release",
                      "audience": "Product and customer-success leaders",
                      "desired_outcome": "Approve one measurable first-30-day onboarding experience",
                      "current_context": "New customers receive inconsistent guidance after the sales handoff, and the team needs one bounded pilot before expanding the workflow.",
                      "constraints": [
                        "No new platform",
                        "One product team",
                        "Six-week pilot"
                      ],
                      "options": [
                        {
                          "title": "Concierge pilot",
                          "description": "A human-led cohort using the current product and a shared checklist."
                        },
                        {
                          "title": "In-product pilot",
                          "description": "A guided workflow implemented inside the current product experience."
                        }
                      ],
                      "trade_offs": [
                        "Reach versus learning depth",
                        "Speed versus automation"
                      ],
                      "evidence": [
                        {
                          "label": "Support handoff review",
                          "evidence": "Author-owned synthesis from recent customer handoffs.",
                          "owner": "Customer success lead",
                          "url": "",
                          "local_reference": "",
                          "checked_date": null,
                          "license": "",
                          "review_status": "unresolved",
                          "claim_ids": []
                        }
                      ],
                      "owner": "Onboarding lead",
                      "next_step": "Run a five-customer concierge pilot and review the evidence",
                      "review_date": "2026-09-30"
                    },
                    "theme": "midnight"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Decision Story Api V1 Stories Decision Brief Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Create Content",
        "description": "Create an editable outline; no request data is retained by the server.",
        "operationId": "create_content_api_v1_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              },
              "examples": {
                "local-no-key": {
                  "summary": "Local No Key",
                  "value": {
                    "topic": "Choose a reliable onboarding direction",
                    "slide_count": 3,
                    "brief": "Align product and customer-success leaders",
                    "use_ai": false,
                    "provider": "local",
                    "slide_configs": []
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Content Api V1 Content Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/content": {
      "post": {
        "tags": [
          "generation"
        ],
        "summary": "Create Content",
        "description": "Create an editable outline; no request data is retained by the server.",
        "operationId": "create_content_api_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Content Api Content Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/presentations": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Export Presentation",
        "description": "Render an isolated export. User content is retained only in that PPTX.",
        "operationId": "export_presentation_api_presentations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPresentationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Export Presentation Api Presentations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bundles": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Export Review Bundle",
        "description": "Export a local PPTX, versioned story, and verifiable receipt as one ZIP.",
        "operationId": "export_review_bundle_api_v1_bundles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryDocumentV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Export Review Bundle Api V1 Bundles Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/validate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Validate Project",
        "operationId": "validate_project_api_v1_projects_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Project Api V1 Projects Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/save": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Save Project",
        "description": "Save editable inputs and assets without requiring a renderable deck.",
        "operationId": "save_project_api_v1_projects_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Save Project Api V1 Projects Save Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresentationPayload": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 90,
            "minLength": 1,
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "maxLength": 110,
            "title": "Subtitle",
            "default": ""
          },
          "theme": {
            "type": "string",
            "enum": [
              "midnight",
              "glacier",
              "ember",
              "forest",
              "royal",
              "sakura"
            ],
            "title": "Theme",
            "default": "midnight"
          },
          "slides": {
            "items": {
              "$ref": "#/components/schemas/SlideContent"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 3,
            "title": "Slides"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/LocalAsset"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Assets"
          },
          "brand_kit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandKit"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations_appendix": {
            "type": "boolean",
            "title": "Citations Appendix",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "slides"
        ],
        "title": "PresentationPayload"
      },
...
    }
  }
}